# TravelerList::clinched_segments -> vector, etc.