@yakra 2023-02-12
a90d66477911bdc60f5cedd62f8c315450b9ee34 saves about 270 MiB on Ubuntu 16.04.