@michihdeu 2023-02-12
So, only 0.03% speed increase on Jim's machine?

> bsdlab 	FreeBSD 13.0 	0.6941 	0.6939 	1.000 	0.03

And 16 Byte x ~360 = ~5,760 Byte RAM saving?