@yakra 2023-02-12
> So, only 0.03% speed increase on Jim's machine?

For 616425bc2fb5df01778f43c2d06ccbe15ae94816, 0.03% speed increase on my FreeBSD machine, same hardware as lab3 & lab4.
You'll see in the same table there's a bit of variation between different machines running the same OS. So noreaster may not be exactly 0.03%, but in the ballpark.
OK, so now I'm curious. Starting a test on nroeaster. :rofl: 1.61% faster! (Table updated.)
Of course, even BiggaTomato, the machine with the most wall time difference, saves a whopping 0.08s. But *it was the right thing to do!*

> And 16 Byte x ~360 = ~5,760 Byte RAM saving?

Yes and no.

**Yes** in that 16 B \* 363 = 5808 B, allocated on the stack, known at compile time. This figure comes from `sizeof(std::unordered_set<HighwaySegment*> - sizeof(std::vector<HighwaySegment*>)` = [40 - 24](https://github.com/yakra/DataProcessing/issues/168) = 16 B.

**No** in that both an [unordered_set](https://cplusplus.com/reference/unordered_set/unordered_set/) and a [vector](https://cplusplus.com/reference/vector/vector/), like many data structures, will have to dynamically allocate some amount of memory, unknown at compile time, on the heap.
* For a simple example, let's look at [Region::allregions](https://github.com/TravelMapping/DataProcessing/blob/cc9da66106a7581974938040418b845dcedaf4b2/siteupdate/cplusplus/classes/Region/Region.h#L51). That's a vector; it's `sizeof` is 24 B.
  * But it has to dynamically allocate an array of `Region` pointers. At 8 B * 510 (including a [dummy](https://github.com/TravelMapping/DataProcessing/blob/1cc2b9570cdfbe6a8cd501c3149712c074de8f4b/siteupdate/cplusplus/siteupdate.cpp#L173-L174)), that's 4080 B.
  * But that's not the whole story. In order to avoid having to reallocate *every time* an item is added, when a vector is resized, it allocates extra memory to allow itself room to grow. In all implementations I've looked at, it doubles in size every time. So its [capacity](https://cplusplus.com/reference/vector/vector/capacity/) is 512 elements.
  * ...making the final size 512*8+24 = 4120 B.
* For a more complex example, consider the `WaypointQuadtree` -- sizeof([all_waypoints](https://github.com/TravelMapping/DataProcessing/blob/1cc2b9570cdfbe6a8cd501c3149712c074de8f4b/siteupdate/cplusplus/siteupdate.cpp#L211-L213)) = 152 B. Not much in and of itself.
  * As `Waypoint*`s are inserted and the quadtree is refined, tens of thousands more `WaypointQuadtree` objects are created. 52,821 of them as of the latest [siteupdate.log](https://travelmapping.net/logs/siteupdate.log), for a total of 8,028,792 B so far.
  * And then, terminal nodes will have a [list](https://cplusplus.com/reference/list/list/) of `Waypoint` pointers. Lists take up more space than vectors; each node contains, in addition to the stored object itself, pointers to the previous & next nodes. 3 pointers * 8 B = 24 B, * 1,238,553 waypoints = 29,725,272 B. Add in those 8 MB for storing the quadtree nodes themselves, and that makes the *real* size of `all_waypoints` 37,754,064 B. **36 MiB!**
* An `unordered_set` will operate somewhere between these 2 extremes. I don't know how exactly they work under the hood (which can differ in different compilers) and don't want to go to the bother of finding out. But they have to manage different [buckets](https://cplusplus.com/reference/unordered_set/unordered_set/bucket/) and arrays of stuff, however that all works; suffice it to say it's more complex than a `vector`. More memory allocated on the heap.
  * Hence mentioning the "hidden cost". Looking at my System Monitor during runs of the old and new executables [on BiggaTomato, it saves about 270 MiB](https://github.com/TravelMapping/DataProcessing/pull/579#issuecomment-1426933955).
  * FreeBSD is a bit different; I can't get as precise a handle on how much RAM is used for each run in [htop](https://htop.dev/). But it does appear to be about the same ballpark.

Other than saving about 270 MiB, the big win here is the speedup.

> btw, ìs there a benefit on still updating Python? Except of having fun and learning....

I'm running with the "continue to update Python in parallel until we decide to officially sunset it" philosophy.
It helps for intelligibility purposes to keep the two versions similar in terms of data structures, functions & algorithms where practicable. Of course, there are enough other big differences throughout the program that, how important *is* this, really?
While still updating Python, what's important is only to keep doing the same datachecks & output the same data -- logs, csvs, .sql file -- the same way.
OTOH, the changes here were so easily done that, hey, why not! Plus I'm always curious to see what the time savings are. That's where the having fun and learning comes in, I guess!